backend: auth: Extract ConfigureTLSContext from headlamp.go
